Remove azcopy as a dependency on traverser
