+++ cmd/copyEnumeratorHelper_test.go
@@ -23,7 +23,6 @@ package cmd
 import (
 	"testing"
 
-	"github.com/Azure/azure-storage-azcopy/v10/azcopy"
 	"github.com/Azure/azure-storage-azcopy/v10/common"
 	"github.com/Azure/azure-storage-azcopy/v10/traverser"
 	"github.com/stretchr/testify/assert"
@@ -34,7 +33,7 @@ func newLocalRes(path string) common.ResourceString {
 }
 
 func newRemoteRes(url string) common.ResourceString {
-	r, err := azcopy.SplitResourceString(url, common.ELocation.Blob())
+	r, err := traverser.SplitResourceString(url, common.ELocation.Blob())
 	if err != nil {
 		panic("can't parse resource string")
 	}
