+++ traverser/clientUtil.go
@@ -1,4 +1,4 @@
-package azcopy
+package traverser
 
 import (
 	"net/http"
